requirements.txt
apipkg==1.5
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.6.16
coverage==4.5.4
dj-database-url==0.5.0
Django==3.0.7
django-fileprovider==0.1.4
django-heroku==0.3.1
django-mysql==3.2.0
django-import-export==2.0.1
django-recaptcha==2.0.6
django-secure==1.0.1
django-toolbox==0.2
django-widget-tweaks==1.4.5
entrypoints==0.3
execnet==1.7.1
flake8==3.7.9
flake8-bugbear==19.8.0
flake8-print==3.1.4
gunicorn==19.9.0
imgkit==1.0.2
importlib-metadata==0.23
mccabe==0.6.1
more-itertools==7.2.0
opencv-python-headless==4.2.0.32
packaging==19.2
pdfkit==0.6.1
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.2.2
pytest-cov==2.8.1
pytest-django==3.6.0
pytest-forked==1.1.3
pytest-xdist==1.30.0
python-dotenv==0.10.3
pytz==2019.3
random-username==1.0.2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
wcwidth==0.1.7
WeasyPrint==51
whitenoise==4.1.4
zipp==0.6.0
pandas==1.0.1
